Update urls and some refactor
