+++ src/cow-react/modules/wallet/api/components/WalletModal/index.tsx
@@ -9,7 +9,7 @@ import { Connector } from '@web3-react/types'
 
 import { AutoColumn } from 'components/Column'
 import { AutoRow } from 'components/Row'
-import { ConnectionType, walletConnectConnection } from 'connection'
+import { ConnectionType, walletConnectConnection } from '@cow/modules/wallet/api/utils/connections'
 import { getConnection, getIsCoinbaseWallet, getIsInjected, getIsMetaMask } from '@cow/modules/wallet/api/utils'
 import { useCallback, useEffect, useState } from 'react'
 import { updateConnectionError } from 'state/connection/reducer'
