+++ src/cow-react/modules/wallet/api/components/WalletModal/options/AlphaWalletOption.tsx
@@ -1,12 +1,11 @@
 import { Connector } from '@web3-react/types'
 import ALPHA_WALLET_ICON_URL from 'assets/images/alphawallet.svg'
-import { ConnectionType, walletConnectConnection } from 'connection'
-import { getConnectionName } from 'connection/utils'
+import { ConnectionType, walletConnectConnection } from '@cow/modules/wallet/api/utils/connections'
+import { getConnectionName, getIsAlphaWallet } from '@cow/modules/wallet/api/utils'
 import { useIsActiveWallet } from 'hooks/useIsActiveWallet'
 
 import Option from '../Option'
 import { useWalletInfo } from '../../../hooks/useWalletInfo'
-import { getIsAlphaWallet } from 'connection/utils'
 
 const BASE_PROPS = {
   color: '#4196FC',
