+++ src/cow-react/modules/wallet/api/components/WalletModal/options/FortmaticOption.tsx
@@ -1,6 +1,6 @@
 import { Connector } from '@web3-react/types'
 import FORTMATIC_ICON_URL from '../../../assets/formatic.png'
-import { ConnectionType, fortmaticConnection } from 'connection'
+import { ConnectionType, fortmaticConnection } from '@cow/modules/wallet/api/utils/connections'
 import { getConnectionName } from '@cow/modules/wallet/api/utils'
 
 import Option from '../Option'
