+++ src/cow-react/modules/wallet/api/components/WalletModal/options/InjectedOption.tsx
@@ -3,7 +3,7 @@ import { Connector } from '@web3-react/types'
 import INJECTED_ICON_URL from '../../../assets/arrow-right.svg'
 import INJECTED_ICON_WHITE_URL from '../../../assets/arrow-right-white.png'
 import METAMASK_ICON_URL from '../../../assets/metamask.png'
-import { ConnectionType, injectedConnection } from 'connection'
+import { ConnectionType, injectedConnection } from '@cow/modules/wallet/api/utils/connections'
 import { getConnectionName } from '@cow/modules/wallet/api/utils'
 
 import Option from '../Option'
