+++ src/cow-react/modules/wallet/api/components/WalletModal/options/WalletConnectOption.tsx
@@ -1,12 +1,12 @@
 import { Connector } from '@web3-react/types'
 import WALLET_CONNECT_ICON_URL from '../../../assets/walletConnectIcon.svg'
-import { ConnectionType, walletConnectConnection } from 'connection'
+import { ConnectionType, walletConnectConnection } from '@cow/modules/wallet/api/utils/connections'
 import { getConnectionName } from '@cow/modules/wallet/api/utils'
 import { useIsActiveWallet } from 'hooks/useIsActiveWallet'
 
 import Option from '../Option'
 import { useWalletInfo } from '../../../hooks/useWalletInfo'
-import { getIsAmbireWallet, getIsZengoWallet, getIsAlphaWallet } from 'connection/utils'
+import { getIsAmbireWallet, getIsZengoWallet, getIsAlphaWallet } from '@cow/modules/wallet/api/utils'
 import { WC_DISABLED_TEXT } from '../../../constants'
 
 const BASE_PROPS = {
