+++ src/state/connection/constants.ts
@@ -1,4 +1,4 @@
-import { ConnectionType } from 'connection'
+import { ConnectionType } from '@cow/modules/wallet/api/utils/connections'
 
 export const BACKFILLABLE_WALLETS = [
   ConnectionType.INJECTED,
